Fix binary-dedupe failures and improve its performance on macOS (#3848)
This PR fixes #3823. 

The minimum fix for this issue is to use `shasum` on macOS vs `sha1sum` on Linux. Once the script is fixed functionally it exhibits unacceptable performance on macOS. The single-threaded performance was also bad on Linux but it was mitigated by using parallel xargs. The slowness may  be attributable to much higher cost of forking a child process and IO due to the presence of  an enterprise security software on macOS. This PR reworks the script such that it is much faster without requiring parallelism, and the run time is similar for both Linux and macOS.

1. eliminate instances of forked process per class file by using bash builtins 
2. compute sha for all files just once
3. build large lists for a few large `rsync` and `rm` calls

On a sample MBP the run time goes down from at least a few minutes (did not wait for completion ) to 25 seconds
On a sample Linux desktop the script run time decreases from over a minute to 3-5 seconds depending on the profile.
 
Signed-off-by: Gera Shegalov <gera@apache.org>
Change-Id: I7eedf7e94a13c6149e9e4b1c6cef644d3f4d36df