CDPD-30259. Create a cloudera 3.2 shim layer for spark-rapids
The Spark320CDHShims class is essentially a copy of Spark320Shims class. The reason a copy had to be made instead of directly inheriting Spark320Shims is that because of the way the project is structured, it is not possible to make spark320cdh module depend on the spark320 module.

Change-Id: Ic909c52a9c79640d593a64ee5193aee3eceed5c6